Module oxidize::dsl

Contains functions wrapping creating matchers in a more pleasant syntax.

Functions

be_false

be_true

contain

empty

equal

expect
Create an expectation with a value that can then be matched against.

greater_than

less_than

match_regex

none

some